core models moved to apps, code isorted
