+++ skaben/access/admin.py
@@ -1,0 +1,5 @@
+from core import models
+from django.contrib import admin
+
+admin.site.register(models.Action)
+admin.site.register(models.UserInput)
