+++ skaben/access/apps.py
@@ -1,5 +1,7 @@
 from django.apps import AppConfig
 
 
-class AccessConfig(AppConfig):
+class AlertConfig(AppConfig):
     name = 'access'
+    verbose_name = 'ALERT: управление уровнями тревоги'
+    default_auto_field = 'django.db.models.BigAutoField'
