+++ skaben/access/views.py
@@ -1,9 +1,10 @@
 from access import serializers
-from core.models import AccessCode, Permission
 from core.views import DynamicAuthMixin
 from rest_framework import viewsets
 from rest_framework.authentication import TokenAuthentication
 from rest_framework.permissions import IsAuthenticated
+
+from .models import AccessCode, Permission
 
 
 class AccessCodeViewSet(viewsets.ModelViewSet, DynamicAuthMixin):
