+++ skaben/actions/admin.py
@@ -1,0 +1,5 @@
+from core import models
+from django.contrib import admin
+
+admin.site.register(models.AlertState)
+admin.site.register(models.AlertCounter)
