+++ skaben/actions/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class ActionsConfig(AppConfig):
+    verbose_name = 'ACTIONS: конфигурация действий'
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'actions'
