+++ skaben/actions/device.py
@@ -1,20 +1,19 @@
 from typing import Optional
 
-from django.conf import settings
-
-from core import models
 from device import serializers
+from device.models import Lock, Simple, Terminal
+from django.conf import settings
 from transport.interfaces import publish_without_producer, send_log
 from transport.rabbitmq import exchanges
 
 DEVICES = {
     'lock': {
         'serializer': serializers.LockSerializer,
-        'model': models.Lock
+        'model': Lock
     },
     'terminal': {
         'serializer': serializers.TerminalMQTTSerializer,
-        'model': models.Terminal
+        'model': Terminal
     },
     # 'rgb': {
     #     'serializer': serializers.SimpleLightSerializer,
