+++ skaben/actions/serializers.py
@@ -1,0 +1,11 @@
+from rest_framework import serializers
+
+from .models import UserInput
+
+
+class UserInputSerializer(serializers.ModelSerializer):
+    """ Serializer for menu item objects """
+
+    class Meta:
+        model = UserInput
+        exclude = ("id", "name",)
