+++ skaben/alert/views.py
@@ -1,5 +1,4 @@
 from alert import serializers
-from core.models import AlertCounter, AlertState
 from core.views import DynamicAuthMixin
 from django.core.exceptions import ObjectDoesNotExist
 from django_filters.rest_framework import DjangoFilterBackend
@@ -8,6 +7,8 @@ from rest_framework.authentication import TokenAuthentication
 from rest_framework.decorators import action
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
+
+from .models import AlertCounter, AlertState
 
 
 class AlertStateViewSet(mixins.ListModelMixin,
