+++ skaben/assets/models/__init__.py
@@ -1,0 +1,2 @@
+from .files import AudioFile, ImageFile, TextFile, VideoFile
+from .games import HackGame
