+++ skaben/assets/views.py
@@ -1,10 +1,12 @@
+from actions.models import UserInput
 from assets import serializers
-from core.models import (AudioFile, HackGame, ImageFile, MenuItem, TextFile,
-                         UserInput, VideoFile, WorkMode)
 from core.views import DynamicAuthMixin
+from device.models import MenuItem, WorkMode
 from rest_framework import viewsets
 from rest_framework.authentication import TokenAuthentication
 from rest_framework.permissions import IsAuthenticated
+
+from .models import AudioFile, HackGame, ImageFile, TextFile, VideoFile
 
 
 class UserInputViewSet(viewsets.ModelViewSet, DynamicAuthMixin):
