--- skaben/core/models/storages.py
@@ -1,1 +1,0 @@
-from assets.storages import *
