+++ skaben/device/models/__init__.py
@@ -1,0 +1,2 @@
+from .config import MenuItem, SimpleConfig, WorkMode
+from .device import Lock, Simple, Terminal
