+++ skaben/eventlog/views.py
@@ -1,4 +1,3 @@
-from core.models import EventLog
 from core.views import DynamicAuthMixin
 from django_filters.rest_framework import DjangoFilterBackend
 from eventlog import serializers
@@ -6,6 +5,8 @@ from rest_framework import viewsets
 from rest_framework.authentication import TokenAuthentication
 from rest_framework.permissions import IsAuthenticated
 from transport.interfaces import send_log, send_websocket
+
+from .models import EventLog
 
 
 class EventLogViewSet(viewsets.ModelViewSet, DynamicAuthMixin):
